branch with backend
I created a branch placing the main and api py files that create the latest DB. Also included are the latest DB which should be used, the facets trial jupyter file and the prediction. The input variable is currently hard coded but this is where it would receive the information from the user and possibly fill in the blanks from the census api.